requirements.txt
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.2
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
distro==1.7.0
Django==3.2.8
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
importlib-metadata==4.8.3
jsonschema==3.2.0
paramiko==2.10.4
pdfkit==1.0.0
phonenumbers==8.12.35
pycparser==2.21
pycryptodome==3.11.0
PyNaCl==1.5.0
PyPDF2==1.26.0
pyrsistent==0.18.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==0.59.0
zipp==3.6.0
